Move files in `client`
