+++ client/src/app/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { HashRouter } from "react-router-dom";
 import { App } from "./components/App";
 import Background from "./components/Background";
-import { WormholeProvider } from "./WormholeProvider";
+import { WormholeProvider } from "./components/WormholeProvider";
 
 ReactDOM.render(
   <WormholeProvider>
